Split MSP into MSP and serial parts
